:bug: fix authorizations in activity module

+++ tooling-activity/src/main/kotlin/com/tooling/activity/service/ActivityService.kt
@@ -3,6 +3,7 @@ package com.tooling.activity.service
 import com.tooling.activity.model.Activity
 import com.tooling.activity.model.ActivityDto
 import com.tooling.activity.repository.ActivityRepository
+import com.tooling.core.exception.InvalidUserGroupException
 import com.tooling.core.service.HeaderReader
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.server.ServerRequest
@@ -16,23 +17,59 @@ class ActivityService(private val activityRepository: ActivityRepository) {
     val ADMIN_GROUP = Flux.just("ADMIN", "ACTIVITY_ADMIN")
   }
 
-  fun insert(activityDto: ActivityDto, request: ServerRequest, groups: Flux<String>) =
-    HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
-      .then(
-        mapActivityDtoToActivity(activityDto, request)
-          .flatMap(activityRepository::insert)
-      )
+  fun insert(activityDto: ActivityDto, request: ServerRequest, userIds: Mono<List<String>>, groups: Flux<String>) =
+    userIds
+      .flatMap { ids ->
+        if (ids.size != 1)
+          throw InvalidUserGroupException("User id is null in header");
 
-  fun update(activityDto: ActivityDto, request: ServerRequest, groups: Flux<String>) =
-    HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
-      .then(
-        mapActivityDtoToActivity(activityDto, request)
-          .flatMap(activityRepository::update)
-      )
+        if (!activityDto.userId.equals(ids[0])) {
+          HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
+            .then(
+              mapActivityDtoToActivity(activityDto, request)
+                .flatMap(activityRepository::insert)
+            )
+        } else {
+          mapActivityDtoToActivity(activityDto, request)
+            .flatMap(activityRepository::insert)
+        }
+      }
 
-  fun delete(tenantId: Mono<String>, id: Mono<String>, groups: Flux<String>) =
-    HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
-      .then(activityRepository.deleteByIdAndTenantId(id, tenantId))
+  fun update(activityDto: ActivityDto, request: ServerRequest, userIds: Mono<List<String>>, groups: Flux<String>) =
+    userIds
+      .flatMap { ids ->
+        if (ids.size != 1)
+          throw InvalidUserGroupException("User id is null in header");
+
+        if (!activityDto.userId.equals(ids[0])) {
+          HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
+            .then(
+              mapActivityDtoToActivity(activityDto, request)
+                .flatMap(activityRepository::update)
+            )
+        } else {
+          mapActivityDtoToActivity(activityDto, request)
+            .flatMap(activityRepository::update)
+        }
+      }
+
+  fun delete(tenantId: Mono<String>, id: Mono<String>, userIds: Mono<List<String>>, groups: Flux<String>) =
+    userIds
+      .flatMap { ids ->
+        if (ids.size != 1)
+          throw InvalidUserGroupException("User id is null in header");
+
+        activityRepository
+          .findById(id)
+          .flatMap { activity ->
+            if (ids[0].equals(activity.userId)) {
+              HeaderReader.oneRuleMatch(groups, ADMIN_GROUP)
+              .then(activityRepository.deleteByIdAndTenantId(id, tenantId))
+            } else {
+              activityRepository.deleteByIdAndTenantId(id, tenantId)
+            }
+          }
+      }
 
   private fun mapActivityDtoToActivity(activityDto: ActivityDto, request: ServerRequest) =
     HeaderReader.getTenantId(request)
